+++ frontend/src/App.tsx
@@ -1,6 +1,8 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Layout from './components/Layout.tsx';
 import HomePage from './pages/HomePage.tsx';
+import LoginPage from './pages/LoginPage.tsx';
+import LoginOut from './pages/LogoutPage.tsx';
 
 export default function () {
     return (
@@ -8,6 +10,8 @@ export default function () {
             <Layout>
                 <Routes>
                     <Route path="/" element={<HomePage />} />
+                    <Route path="/login" element={<LoginPage />} />
+                    <Route path="/logout" element={<LoginOut />} />
                 </Routes>
             </Layout>
         </Router>